new menu background (simple), new button, new chained Action items 	new file:   StrategyGame/assets/Button.png 	new file:   StrategyGame/assets/Button.psd 	modified:   StrategyGame/assets/intro.png 	modified:   StrategyGame/assets/intro.psd 	new file:   StrategyGame/assets/menubkg.png 	new file:   StrategyGame/assets/menunkg.psd New Images 	modified:   StrategyGame/src/action.cpp 	modified:   StrategyGame/src/action.h Modified to return vector/list of new Actions after it's done. 	modified:   StrategyGame/src/actionintro.cpp Covers the intro screen 	new file:   StrategyGame/src/actionselectserver.cpp 	new file:   StrategyGame/src/actionselectserver.h Pick server (click it away) 	modified:   StrategyGame/src/assetmgr.cpp 	modified:   StrategyGame/src/assetmgr.h Added x and y for grid size 	modified:   StrategyGame/src/game.cpp Changed the Action loop f

+++ StrategyGame/src/unit.h
@@ -41,15 +41,15 @@ public:
 	std::string GetJson();
 	void Draw(SDL_Rect dest);
 
-	inline int GetX() const { return (int) x+.5f; }
-	inline int GetY() const { return (int) y+.5f; }
+	inline int GetX() const { return (int) (x+.5f); }
+	inline int GetY() const { return (int) (y+.5f); }
 	//These return the float positions.
 	inline float GetXF() const { return x; }
 	inline float GetYF() const { return y; }
 
 	inline int GetSpeed() const { return speed; }
 	inline void Move(int dx, int dy) {
-		x = dx; y = dy;
+		x = (float) dx; y = (float) dy;
 	}
 	//This is for animations, to partial move the character.
 	inline void MoveF(float dx, float dy) {
